refactor: change folder structure and add `OAuthLoginButtons.tsx`
