+++ src/app/(auth)/login/page.tsx
@@ -1,8 +1,7 @@
-import { Button, Divider } from 'antd';
-import GoogleIcon from '@/shared/container/Icon/GoogleIcon';
 import Link from 'next/link';
-import LoginForm from '../_container/Form/LoginForm';
-import LoginFormContainer from '../_container/LoginFormContainer';
+import LoginFormContainer from './_container/LoginFormContainer';
+import LoginForm from './_container/Form/LoginForm';
+import OAuthLoginButtons from '../_container/OAuthLoginButtons';
 
 export default function LoginPage() {
   return (
@@ -13,16 +12,7 @@ export default function LoginPage() {
         <LoginForm />
       </LoginFormContainer>
 
-      <Divider plain className="text-caption-2">
-        Or continue with
-      </Divider>
-
-      <Button
-        className="w-full flex items-center justify-center gap-2 bg-ny-primary-100 text-ny-primary-500"
-        size="large"
-        type="primary">
-        <GoogleIcon /> Continue with Google
-      </Button>
+      <OAuthLoginButtons withDivider dividerText="Or continue with" />
 
       <div className="text-caption-1 mt-5 text-center w-full">
         Don&apos;t have an account?{' '}
